requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyDrive==1.3.1
pyparsing==3.2.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
uritemplate==4.1.1
urllib3==2.2.3
